Allow the user to extract a specific glyph
I wanted to extract an emoji from a vector font at high resolution, and
didn't want to dump the every glyph, so I wrote this.

I was able to test that the SVG code correctly tried to extract the
right glyph from an SVG font, but I was unable to get the extraction
itself to work correctly to verify.